+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/common/blockentities/PeripheralBlockEntity.kt
@@ -4,7 +4,6 @@ import dan200.computercraft.api.peripheral.IComputerAccess
 import net.minecraft.core.BlockPos
 import net.minecraft.core.Direction
 import net.minecraft.nbt.CompoundTag
-import net.minecraft.server.level.ServerLevel
 import net.minecraft.world.entity.player.Player
 import net.minecraft.world.level.block.entity.BlockEntity
 import net.minecraft.world.level.block.entity.BlockEntityType
@@ -13,6 +12,7 @@ import site.siredvin.peripheralium.api.blockentities.IOwnedBlockEntity
 import site.siredvin.peripheralium.api.peripheral.IOwnedPeripheral
 import site.siredvin.peripheralium.api.peripheral.IPeripheralProvider
 import site.siredvin.peripheralium.api.peripheral.IPeripheralTileEntity
+import java.util.UUID
 
 abstract class PeripheralBlockEntity<T : IOwnedPeripheral<*>>(
     blockEntityType: BlockEntityType<*>,
@@ -24,10 +24,10 @@ abstract class PeripheralBlockEntity<T : IOwnedPeripheral<*>>(
         protected set
 
     protected var peripheral: T? = null
-    protected var owningPlayer: Player? = null
+    protected var ownerPlayerUUID: UUID? = null
     override var player: Player?
-        get() = owningPlayer
-        set(value) { owningPlayer = value }
+        get() = ownerPlayerUUID?.let { level?.getPlayerByUUID(it) }
+        set(value) { ownerPlayerUUID = value?.uuid }
 
     val connectedComputers: List<IComputerAccess>
         get() = if (peripheral == null) emptyList() else peripheral!!.connectedComputers
@@ -54,15 +54,15 @@ abstract class PeripheralBlockEntity<T : IOwnedPeripheral<*>>(
         if (!peripheralSettings.isEmpty) {
             compound.put(PERIPHERAL_DATA_TAG, peripheralSettings)
         }
-        if (owningPlayer != null && owningPlayer!!.gameProfile.id != null) {
-            compound.putUUID(OWNER_PROFILE_TAG, owningPlayer!!.gameProfile.id)
+        if (ownerPlayerUUID != null) {
+            compound.putUUID(OWNER_PROFILE_TAG, ownerPlayerUUID!!)
         }
     }
 
     override fun load(compound: CompoundTag) {
         if (compound.contains(PERIPHERAL_DATA_TAG)) peripheralSettings = compound.getCompound(PERIPHERAL_DATA_TAG)
-        if (compound.contains(OWNER_PROFILE_TAG) && level != null && !level!!.isClientSide) {
-            owningPlayer = (level!! as ServerLevel).getPlayerByUUID(compound.getUUID(OWNER_PROFILE_TAG))
+        if (compound.contains(OWNER_PROFILE_TAG)) {
+            ownerPlayerUUID = compound.getUUID(OWNER_PROFILE_TAG)
         }
         super.load(compound)
     }